bug fix
solves the bug described in the previous commit.
pickled jobs contains upward references in there object tree:
. job.job_sample.parent_job
. job.sampler
after unpicking the job these upward references refer to the objects
that existed at the time of pickling, not to the current objects.
As a consequence the code in NeighbouringJobInfo.__init__():
    job2 = job_sample.parent_job.sampler.jobs[jobid2]
returns the Job as it existed when it was pickled, which obviously does
not contain the current timestamp
After the upward references are corrected the problem is gone.